pyproject.toml
[tool.poetry]
name = "petproject"
version = "0.1.0"
description = ""
authors = ["Denis <denis.zinovkin234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
environs = "^14.2.0"
django = "^5.2.4"
psycopg2-binary = ">=2.9,<3.0"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
djangorestframework = "^3.16.0"
celery = "^5.5.3"
redis = "^6.2.0"
aiogram = "^3.21.0"
ollama = "^0.5.2"
httpx = "^0.28.1"
wikijs-pyclient = "^1.0.1"
chromadb = "^1.0.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
